Use new `depends.docker-image` support in `build.yml` for image cache packages.
This means that these packages are no longer a special case requiring
supporting `Makefile` scaffolding to build. In turn this means that CI can
build and push these images alongside all the others.

To avoid skew, CI also checks that the `build.yml` and `Dockerfile` are
identical apart from the expected difference in the name. Extend the existing
`check-hashes` job into a more generic `lint`.

Signed-off-by: Ian Campbell <ijc@docker.com>