CI: Retag images before build since they maybe unsigned so far
Images are only signed when they are pushed to hub. An image which has only
just been built by this CI run will not yet have been pushed (since that
happens last) and hence will not be signed.

To avoid this and allow images to be built with content trust (for all other
images) retag built images and update the yml on the fly to use the retagged
versions. Since the hub org used for retagging is not listed in `trust.org` the
build wont care that they haven't been signed.

We retag all the local packages even if they are not new in this build (i.e.
they might already be on hub and be signed), it's hard to determine when this
is the case and not really necessary in this context. The content trust will
have been checked on pull (done as part of `linuxkit pkg build`).

Signed-off-by: Ian Campbell <ijc@docker.com>