Generic make targets
Depending on the options used you may or may not be building `kube-master.iso`
etc. Given the targets generic/phony names instead.

Also the dependencies for `kube-master.iso` were not complete so it wouldn't
always rebuild when you expected it too (`kube-master.iso is already up to
date`) so this change side-steps that too.

Signed-off-by: Ian Campbell <ijc@docker.com>