Update alpine base image in update-linuxkit-hashes.sh
The hash is pulled from the versions.x86_64 (which is used for the multi-arch
manifest too). Since we are updating local packages we then run update_hashes
after committing. It needs to be after commiting else we would get a `-dirty`
suffix on the previous hash instead of the new hash.

Signed-off-by: Ian Campbell <ijc@docker.com>