Portable `make update-hashes`.
MacOS lacks `xargs`' `--no-run-if-empty` option. Since the `git grep` always
matches at least one file it's not really needed.

However since the files which need patching are just those under yml/ which are
small enough in number and size that we can just unconditionally apply the
update to the all.

Fixes #52.

Signed-off-by: Ian Campbell <ijc@docker.com>