test: Retry intra-pod networking test.
It seems that pod DNS can take a while to settle and trying to quickly after
the pods are up can fail with DNS lookup failure.

So switch to a model where we retry for up to 60 seconds.

Also switch to curl for consistency with the external test, this also involves
an `apk add` which is a nice test of external connectivity.

Signed-off-by: Ian Campbell <ijc@docker.com>