bridge: Drop `set -x` and add metadata-activated debug knob.
The `set -x` merely adds a `: cat` to the logs, which is useless so drop it.
Instead replace with a metadata key which causes the full config to be printed
to the console.

Using `grep .` means each line is prefixed with the filename and also handles
the case where the file lacks a final newline which e.g. a simple `cat` would
not.

Signed-off-by: Ian Campbell <ijc@docker.com>