Update to Kubernetes v1.10.0.
This also required updating some image versions in the caches:

    kube_dns_version: 1.14.7 → 1.14.8
    pause_version:    3.0    → 3.1
    etcd_version:     3.1.11 → 3.1.12

From https://kubernetes.io/docs/imported/release/notes/#before-upgrading:

    [action-required] The Container Runtime Interface (CRI) version has increased from v1alpha1 to v1alpha2.
      Runtimes implementing the CRI will need to update to the new version, which configures container
      namespaces using an enumeration rather than booleans. (#58973, @verb)

Thus we must update to a newer cri-containerd in lock-step. However this is
made more complicated because cri-containerd is no longer a standalone daemon
but has become a containerd plugin, which will require some rearchitecting (and
which is so far only part of containerd v1.1-rc).  Luckily the version right
before standalone mode was deleted is available in a branch and supports the
required API version so switch to that here as a stop gap measure.

Note that the naming has changed kubernetes-incubator/cri-containerd →
containerd/cri-containerd → containerd/cri with the branch we are using being
at the middle step so update the naming and paths appropriately.

The command line options have changed a bit, so adjust.

Signed-off-by: Ian Campbell <ijc@docker.com>